Improvement to Login
Redirects to group page and restricts access for all other users!

To test it:

username/password

karl/karl - Manager
pedro/pedro - Coach
zoe/zoe - Parent

everyone gets a different view and can not access the other pages.
Other pages can be accessed if for example karl is added to parent.
This will karl permission to access the manager pages and the parent
pages