MANY MANY Changes !
All the previous data from the DB has been wiped. All the data that is
currently within the application is built only using the tools we built
(The feeling is really cool :D). All the major features of the system
have ben tested to function by me but not the minor ones e.g.: some
buttons still need fixing. Introduced some consistency to the
referencing of pages within the application. Some more changes to the
model and to the corresponding forms.